researchpilot/requirements.txt
google-generativeai>=0.5.0
arxiv>=2.1.0
requests>=2.31.0
beautifulsoup4>=4.12.0
scholarly>=1.7.11
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
sqlalchemy>=2.0.0
alembic>=1.13.0
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.6.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
httpx>=0.27.0
rich>=13.7.0
typer>=0.12.0
numpy>=1.26.0
pandas>=2.2.0
PyPDF2>=3.0.0
pypdf>=4.1.0
tenacity>=8.2.0
loguru>=0.7.0
python-multipart>=0.0.9
jinja2>=3.1.0
